Change KeyBindData_v1 format to handle unknown actions better
This modifies how the game loads in new keybindings and makes sure unknown actions get discarded.

There is an issue however, switching between older and newer version of the game might cause you to lose bound action. This is because the older game will discard an unknown action, however upon running the newer game, they action will be rebound if key is not already assigned to a different action.